Moved PluginEnvironmentHelpers
