+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/StaticHelpers/PluginEnvironmentHelpers.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GraphQLAPI\GraphQLAPI\PluginManagement;
+namespace GraphQLAPI\GraphQLAPI\StaticHelpers;
 
 class PluginEnvironmentHelpers
 {
